# api-documentation.yaml
openapi: 3.0.3
info:
  title: PLOMES Route Optimizer API
  description: |
    API for the PLOMES Route Optimization system that integrates with Ploome CRM
    to manage customers, geocoding, and route optimization for Ciara Máquinas.

    ## Features
    - JWT-based authentication
    - Customer data synchronization with Ploome CRM
    - Real-time geocoding services
    - Route optimization algorithms
    - Statistics and analytics dashboard

    ## Database
    - **Backend**: Supabase PostgreSQL (persistent storage)
    - **Current Data**: 2,246+ active customers with geocoding
    - **Integration**: Live Ploome CRM synchronization

  version: 1.0.0
  contact:
    name: PLOMES Route Optimizer
    email: gustavo.canuto@ciaramaquinas.com.br
  license:
    name: Private
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://plomes-route-optimizer.vercel.app
    description: Production server (if deployed)
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: User Login
      description: Authenticate user and receive JWT token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "gustavo.canuto@ciaramaquinas.com.br"
                password:
                  type: string
                  example: "ciara123@"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    description: JWT authentication token
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/verify:
    get:
      summary: Verify Token
      description: Verify if JWT token is valid and return user information
      tags:
        - Authentication
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Token is valid"
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "gustavo.canuto@ciaramaquinas.com.br"
        name:
          type: string
          example: "Gustavo Canuto"
        lastLogin:
          type: string
          format: date-time
          example: "2025-09-24T23:15:41.638+00:00"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
...